requirements.txt
python-dotenv==0.15.0
celery==5.0.5
redis==3.5.3
pytest==6.2.1
Flask==1.1.2
mediacloud==3.12.1
gunicorn==20.0.4
requests==2.25.1
